replace call to wipe with burn event
The SaiTub_call_wipe table does not cover wipes that happen in the context of a `shut` (CDP being closed). So we replace it with the table of SAI burn events. Because we're doing a (INNER) JOIN with the MKR burns, only SAI burns related to wipes/shuts are returned.